Fix bug in Getting Started section
Without a fix, this server will validate a request that includes valid values transmitted via another vector (e.g. query string) but only use parameters from the request body. `check` will reject an invalid value for a parameter found on the body, but not an empty value. This is a bug, and repeating this mistake may create security vulnerabilities if programmers expect `req.body` to contain the parameters described by the array of validation middlewares preceding the request handler.